feat(science): add link

+++ engine-ui/packages/console-ui/src/webapps/science/views/workbench/mainBench/panelGroup/notebookGroup/index.js
@@ -10,6 +10,7 @@ import DefaultNotebookView from '../../default/defaultNoteBookView';
 import * as tabActions from '../../../../../actions/base/tab';
 import { siderBarType } from '../../../../../consts';
 import * as commActions from '../../../../../actions/base';
+import { checkAndcloseTabs } from '../helper';
 
 const TabPane = Tabs.TabPane;
 
@@ -31,27 +32,42 @@ class NoteBookGroup extends React.Component {
     switchTab (key) {
         this.props.setCurrentTab(siderBarType.notebook, key)
     }
-    closeTabs (type) {
-        const { currentTabIndex } = this.props;
+    async closeTabs (type) {
+        const { tabs = [], currentTabIndex } = this.props;
         switch (type) {
             case 'ALL': {
-                this.props.deleteAllTab(siderBarType.notebook);
+                let isChecked = await checkAndcloseTabs(tabs, tabs.map((tab) => {
+                    return tab.id
+                }));
+                if (isChecked) {
+                    this.props.deleteAllTab(siderBarType.notebook);
+                }
                 break;
             }
             case 'OHTERS': {
-                this.props.deleteOtherTab(siderBarType.notebook, currentTabIndex);
+                let isChecked = await checkAndcloseTabs(tabs, tabs.map((tab) => {
+                    return tab.id
+                }).filter((id) => {
+                    return id != currentTabIndex
+                }));
+                if (isChecked) {
+                    this.props.deleteOtherTab(siderBarType.notebook, currentTabIndex);
+                }
                 break;
             }
         }
     }
-    closeTab (tabId) {
+    async closeTab (tabId) {
         const { tabs = [], currentTabIndex } = this.props;
-        if (currentTabIndex == tabId && tabs.length > 1) {
-            this.switchTab(tabs.filter((tab) => {
-                return tab.id != currentTabIndex
-            }).pop().id);
+        let isChecked = await checkAndcloseTabs(tabs, [tabId]);
+        if (isChecked) {
+            if (currentTabIndex == tabId && tabs.length > 1) {
+                this.switchTab(tabs.filter((tab) => {
+                    return tab.id != currentTabIndex
+                }).pop().id);
+            }
+            this.props.deleteTab(siderBarType.notebook, tabId);
         }
-        this.props.deleteTab(siderBarType.notebook, tabId);
     }
     render () {
         const { tabs = [], currentTabIndex } = this.props;
